refactor: dbmanager instance now in plugins
