+++ plugins/database/plugin.go
@@ -1,11 +1,13 @@
-// Package databaseplugin is a plugin that manages the badger database (e.g. garbage collection).
-package databaseplugin
+// Package database is a plugin that manages the badger database (e.g. garbage collection).
+package database
 
 import (
+	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 	"github.com/iotaledger/wasp/packages/database/dbmanager"
 	"github.com/iotaledger/wasp/packages/parameters"
 
 	"github.com/iotaledger/hive.go/daemon"
+	"github.com/iotaledger/hive.go/kvstore"
 	"github.com/iotaledger/hive.go/logger"
 	"github.com/iotaledger/hive.go/node"
 )
@@ -14,6 +16,8 @@ const pluginName = "Database"
 
 var log *logger.Logger
 
+var dbm *dbmanager.DBManager
+
 // Init is an entry point for the plugin.
 func Init() *node.Plugin {
 	return node.NewPlugin(pluginName, node.Enabled, configure, run)
@@ -21,13 +25,13 @@ func Init() *node.Plugin {
 
 func configure(_ *node.Plugin) {
 	log = logger.NewLogger(pluginName)
-	dbmanager.CreateInstance(logger.NewLogger("dbmanager"), parameters.GetBool(parameters.DatabaseInMemory))
+	dbm = dbmanager.NewDBManager(logger.NewLogger("dbmanager"), parameters.GetBool(parameters.DatabaseInMemory))
 
 	// we open the database in the configure, so we must also make sure it's closed here
 	err := daemon.BackgroundWorker(pluginName, func(shutdownSignal <-chan struct{}) {
 		<-shutdownSignal
 		log.Infof("syncing database to disk...")
-		dbmanager.Instance.Close()
+		dbm.Close()
 		log.Infof("syncing database to disk... done")
 	}, parameters.PriorityDatabase)
 	if err != nil {
@@ -36,8 +40,20 @@ func configure(_ *node.Plugin) {
 }
 
 func run(_ *node.Plugin) {
-	err := daemon.BackgroundWorker(pluginName+"[GC]", dbmanager.Instance.RunGC, parameters.PriorityBadgerGarbageCollection)
+	err := daemon.BackgroundWorker(pluginName+"[GC]", dbm.RunGC, parameters.PriorityBadgerGarbageCollection)
 	if err != nil {
 		log.Errorf("failed to start as daemon: %s", err)
 	}
+}
+
+func GetRegistryKVStore() kvstore.KVStore {
+	return dbm.GetRegistryKVStore()
+}
+
+func GetOrCreateKVStore(chainID *ledgerstate.AliasAddress, dedicatedDbInstance bool) kvstore.KVStore {
+	return dbm.GetOrCreateKVStore(chainID, dedicatedDbInstance)
+}
+
+func GetKVStore(chainID *ledgerstate.AliasAddress) kvstore.KVStore {
+	return dbm.GetKVStore(chainID)
 }
